Renamed conditionCheckers to filterSelections
