+++ src/framework/filterControl/_index.scss
@@ -1,7 +1,7 @@
 @import 'filterControl';
 
-@import 'conditionCheckers/conditionCheckerForm';
-@import 'conditionCheckers/conditionCheckerList';
+@import 'filterSelections/conditionCheckerForm';
+@import 'filterSelections/conditionCheckerList';
 @import 'filterDropdown/filterDropdown';
 @import 'filterOptions/FilterOptionList';
 @import 'forms/filterForm';
