+++ src/framework/filterControl/filterSelections/_conditionCheckerList.scss
